test: PlayingState 클래스 테스트 코드 작성
